+++ theme/src/client/composables/index.ts
@@ -1,13 +1,20 @@
-export * from './dark-mode.js'
-export * from './scroll-promise.js'
 export * from './theme-data.js'
+export * from './dark-mode.js'
+export * from './data.js'
+export * from './scroll-promise.js'
+
 export * from './sidebar.js'
 export * from './aside.js'
-export * from './page.js'
+export * from './outline.js'
+
+export * from './prev-next.js'
+export * from './edit-link.js'
+export * from './latest-updated.js'
+export * from './contributors.js'
+
 export * from './blog.js'
 export * from './tag-colors.js'
+
 export * from './locale.js'
 export * from './route-query.js'
 export * from './watermark.js'
-export * from './data.js'
-export * from './outline.js'
